docs: create README.md for the project
A README.md file has been added to the project. This file provides a
brief description of the project, installation instructions, usage
examples, a list of available commands, instructions for running tests,
guidelines for contributing, and license information. This will help
users and contributors understand how to use and contribute to the
project.